fix: silence DB logs by default (#324)
This commit fixes a major security issue of SQL queries being leaked to
the console. This is because by default the GORM logger is set to log
when a query errors (i.e. a record is not found). We are turning this
down to silent and removing the check on startup to log all queries for
local development.

Fixes #323