fix(clock): disable start on render when clock is paused (#786)
Fixes an issue where the clock could not be paused once a level had advanced. The issue was that the `startOnRender` prop and `shouldStart` state was being set to `true` when the level advanced, however it was never being set back to `false`, so if you paused the clock, a new render would happen, and the effect watching `startOnRender` would run and see that it is `true` and would restart the clock.

Fixes #785